WIP some more

+++ production/gateway/src/controllers/user.controller.ts
@@ -10,22 +10,38 @@ import {
   Post,
   Put,
   Query,
+  Req,
   Res,
+  UploadedFiles,
+  UseGuards,
   UseInterceptors,
   UsePipes,
   ValidationPipe,
 } from '@nestjs/common';
+import { FilesInterceptor } from '@nestjs/platform-express';
 import { ApiOperation, ApiTags } from '@nestjs/swagger';
 import { Response } from 'express';
+import { AuthGuard } from 'src/guards/jwt.guard';
+import { RegistrationGuard } from 'src/guards/registration.guard';
 import { isError } from 'src/interfaces/error.interface';
-import { CreateUserDto, UsersResponseDto } from 'src/interfaces/user.interface';
+import {
+  CreateUserDto,
+  UserRole,
+  UsersLoggedResponseDto,
+  UsersResponseDto,
+} from 'src/interfaces/user.interface';
 import { IDParams } from 'src/interfaces/validators/id';
 import { PaginationParams } from 'src/interfaces/validators/pagination';
+import { FileService } from 'src/services/file.service';
 import { UserService } from 'src/services/user.service';
 
 @Controller('users')
+@UseGuards(AuthGuard)
 export class UserController {
-  constructor(private readonly userService: UserService) {}
+  constructor(
+    private readonly userService: UserService,
+    private readonly fileService: FileService,
+  ) {}
 
   @Get('/health')
   @ApiTags('Health')
@@ -39,16 +55,26 @@ export class UserController {
   @UsePipes(new ValidationPipe({ transform: true, whitelist: true }))
   @UseInterceptors(ClassSerializerInterceptor)
   async getUsers(@Query() query: PaginationParams) {
-    const users = await this.userService.getUsers(
+    const response = await this.userService.getUsers(
       query.offset,
       query.limit,
       query.find,
     );
 
     return {
-      items: users.map((user) => new UsersResponseDto(user)),
-      total: users.length,
+      items: response.users.map((user) => new UsersResponseDto(user)),
+      total: response.total,
     };
+  }
+
+  @Get('/me')
+  @ApiOperation({ summary: 'Get logged user' })
+  @ApiTags('Users')
+  @UsePipes(new ValidationPipe({ transform: true, whitelist: true }))
+  @UseInterceptors(ClassSerializerInterceptor)
+  async getLoggedUser(@Req() req) {
+    const username = req.user.username;
+    return new UsersLoggedResponseDto(await this.userService.getUser(username));
   }
 
   @Get('/:id')
@@ -73,13 +99,6 @@ export class UserController {
     };
   }
 
-  @Get('/me')
-  @ApiOperation({ summary: 'Get logged user' })
-  @ApiTags('Users')
-  async getLoggedUser() {
-    return {};
-  }
-
   @Post('/')
   @ApiOperation({ summary: 'Create user' })
   @ApiTags('Users')
@@ -90,6 +109,7 @@ export class UserController {
     }),
   )
   @UseInterceptors(ClassSerializerInterceptor)
+  @UseGuards(RegistrationGuard)
   async createUser(@Body() body: CreateUserDto) {
     try {
       const user = await this.userService.createUser(body);
@@ -115,22 +135,45 @@ export class UserController {
   @Put('/:id')
   @ApiOperation({ summary: 'Update user by id' })
   @ApiTags('Users')
-  async updateUser() {
+  @UseInterceptors(FilesInterceptor('file'))
+  async updateUser(
+    @Param() params: IDParams,
+    @Body() body: any,
+    @UploadedFiles() file: Express.Multer.File,
+  ) {
+    if (file) {
+      const avatar = await this.fileService.saveAvatarFile(params.id, file);
+      body.avatar = avatar.split('/').pop();
+    }
+    console.log(body);
+
+    const user = await this.userService.updateUser(params.id, body);
+
+    console.log(user);
+
     return {};
   }
-
-  // @Delete('/')
-  // @ApiOperation({ summary: 'Delete logged user' })
-  // @ApiTags('Users')
-  // async removeUser() {
-  //   return {};
-  // }
 
   @Delete('/:id')
   @ApiOperation({ summary: 'Delete user by id' })
   @ApiTags('Users')
   @UsePipes(new ValidationPipe({ transform: true, whitelist: true }))
-  async removeUserById(@Param() params: IDParams, @Res() res: Response) {
+  async removeUserById(
+    @Param() params: IDParams,
+    @Res() res: Response,
+    @Req() req,
+  ) {
+    if (req.user.username !== params.id || req.user.role !== UserRole.ADMIN) {
+      res
+        .status(HttpStatus.FORBIDDEN)
+        .json({
+          statusCode: HttpStatus.FORBIDDEN,
+          error: 'Forbidden',
+        })
+        .end();
+      return;
+    }
+
     const success = await this.userService.removeUser(params.id);
 
     if (!success) {
